# improve performance for findValues