Updates the RoaringTagIndex.findValues method to loop
over the set of values for a key rather than the set
of matching items in some cases. Depending on the number
of matches, this can often be much faster.

Performance difference was tested using the data from
one of the large prod clusters. The first is just dumping
all values for a given key with no query restriction.
This is the ideal case for this change as before it would
require analyzing all items in the index with that key.

The second test was to dump all values for a key in a
given region. For some values, like the region key it can
be close to the worst case for looping across the values
because keys like region, zone, asg, and node are specific
to a region. The worst case for `hasNonEmptyIntersection`
is when both bitmaps have a lot of values and the result
is empty. While those cases do get a bit worse with this
change, overall it seems worthwhile for improvements to
more common cases.

**Results**

`/api/v1/tags/$key`

|             key | before (ms) |  after (ms) |
|-----------------|-------------|-------------|
|          nf.app |       34.34 |        0.32 |
|      nf.cluster |       89.36 |        0.50 |
|          nf.asg |       96.59 |        0.74 |
|         nf.node |       97.88 |        8.91 |
|      nf.account |        0.96 |        0.01 |
|       nf.region |        0.11 |        0.01 |
|         nf.zone |        0.84 |        0.01 |
|            name |       76.66 |        0.29 |

`/api/v1/tags/$key?q=nf.region,us-east-1,:eq`

|             key | before (ms) |  after (ms) |
|-----------------|-------------|-------------|
|          nf.app |       18.44 |        2.90 |
|      nf.cluster |       46.78 |        9.79 |
|          nf.asg |       47.12 |       57.23 |
|         nf.node |       70.41 |       78.08 |
|      nf.account |        0.84 |        0.19 |
|       nf.region |        0.16 |       15.94 |
|         nf.zone |        0.24 |       31.48 |
|            name |       53.75 |        2.03 |